Added patch step in Dockerfile for nvs_partition_gen.py
The nvs_partition_gen.py file had a small 3 character bug in it. The
older versions of esp-idf nvs support had a limit of 1984 bytes for
string values in nvs. This is no longer the case as all of the
documentation now says it supports 4000 bytes for string values.
However, while the documentation says it should support the 4000 bytes,
the script itself is still enforcing the old limit even though the new
limit exists as a variable in the script. I have corrected this in this
file and copied it to the build environment to patch this specific file.